- Added support for 2D plot curve labels. - Updated the user manual accordingly.
git-svn-id: http://svn.berlios.de/svnroot/repos/qtiplot/trunk@1271 5a6a7de5-fb12-0410-b871-c33778c25c60

+++ qtiplot/src/plot2D/Spectrogram.cpp
@@ -516,7 +516,7 @@ void Spectrogram::setLabelsWhiteOut(bool whiteOut)
 	}
 }
 
-void Spectrogram::setLabelsOffset(int x, int y)
+void Spectrogram::setLabelsOffset(double x, double y)
 {
     if (x == d_labels_x_offset && y == d_labels_y_offset)
         return;
